feat: implemented search (without functionality)
When the search is focused the border radius is removed, the width is set to full and the margin top will be set to 0. Also, the view that contain the categories is hidden and repaced with the search results. Each result contain an artwork (if the artwork is not provided, the default is used) with the title and author.

Co-authored-by: Martín Díaz <43992893+Pegasus8@users.noreply.github.com>