Simplified logic of the package `podcasts`
So, the structures `Podcast` and `Episode` have been moved to the
package `models` (as commited before). Also, the method previously named
`GetPodcast` has been renamed to `GetPodcastData`, and now it
aditionally returns the entire feed. That last move will help us to
avoid the double requests that we've been doing until now to update the
episodes of a subscribed podcast (see the old package `queue` where they
were used before), alongside with the refactoring of function
(previously was a method of the removed structure `Podcast`)
`GetEpisodes`, which now will receive the feed (`*gofeed.Feed`)
as an argument to execute the task of parsing the episodes in their
structs.