Handlers moved and refactored
A lot of methods and functions have been moved to those handlers, so
well, there is a lot of changes here. The vast majority of them are
related with the replacement of the `Database` structure (and its methods) directly with
the structure of `gorm` that manages the database (gorm.DB). And, as a
consequence of the previous committed changes, a lot of logic of other
sides is here now.