Better implementation of handlers
The idea of how to do it is from Miniflux. See this directory:
https://github.com/miniflux/v2/tree/master/api

This implementation is better because there is no global variable
storing the pointer to the instance of the database, keeping things a
bit more organized. Also, this approach is better for testing, since an
x number of instances can be created without cracking our head with
mutexes at time of accessing to the database.