fix(api): resolved data duplication on subscriptions
The problem was when checking the existence of a feed on the db. Since some URLs provided are “masks” of the real feed's URL (basically they do a redirect), and since we were storing the resolved URL instead of the given one, the comparison between them caused in some cases a false positive. As a consequence, using the API with a "masked" feed always leaded to the insertion of it into the db (hence the problem).
To solve that, before check if the feed is already on the database, we resolve the information of it and then, with the real feed URL, we do the comparison with the feeds that are in the database. This will slow down the API of subscriptions, but considering that it won't be frequently requested, I think that it's worth, at least for now.