refactor(api): discrimination of errors on unsubscribe
If there's an error when trying to change the subscription of a podcast to `false`,  the error will be checked to see if is of type `gorm.ErrRecordNotFound` or another. If is of that type, the HTTP status code returned will be 400 Bad Request, since the ID provided by the client was not found in the database (hence the error). On the other hand, if that is not the case, the HTTP status code returned will be 500, because is something unexpected.