test(api): removed check of `Content-Length` headers
`http.Response.ContentLength` seems to be a value no reliable when testing, since there are responses that should contain the header `Content-Length` with the value, but apparently is not captured for some reason. The header don't even appear in the response recorded, so I don't have an idea about what is going on.