fix(api): reject the request when the ID does not exist
The usage of a non-existent podcast ID on `Manager.GetEpisodesHandler` was not rejecting the request as it's supposed to do due to an incorrect way of detecting that behavior (basically, we were expecting that `gorm` throw us an error, something that doesn't happen in this situation). That has been fixed by simply counting the amount of episodes returned, if it's 0, then the ID does not exist.

While this solution is effective (at least for now), there are some situations where it can throw a false positive. For example, If the feed of the podcast has not been updated for the first time (which is supposed to happen at the time of adding it, but right now that is working like that).